requirements.txt
google-generativeai>=0.3.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
